requirements.txt
djangorestframework
django-filter
psycopg2
drf-nested-routers
Django~=3.2.8
